multi: update invalid sig errors to tlv enriched errors
This commit is a placeholder to help us switch all the link errors
over to our new kind of error.

For invalid sig errors, there are 3 options:
1. Use the "erroneous field" tlv type (this omits extra useful data)
2. Create a new CodedError (this omits extra useful data)
3. New tlv type (as is here), works but is a bit incongruous with the
  classification that we've been going with so far in this PR

A possible direction here is to look into the concept of "nested" TLVs,
where we embed TLVs inside of our TLVs because we heard you like TLVs.
This would allow us to extend the erroneous field or codedError tlv
type to contain contextual information about our errors while sticking
with classification thus far.