multi: add encrypted data and metadata to onion payload / hops
This commit adds the encrypted_data and blinding_point tlvs to
the known set of even tlvs for the onion payload. These TLVs are
added in two places (the onion payload and hop struct) because
lnd uses the same set of TLV types for both structs (and they
inherently represent the same thing).

Note: in some places, unit tests intentionally mimic the style
of older tests, so as to be more consistently readable.