multi: add blinded paths to query routes api
When introducing blinded paths to our existing apis, there are a
few already-existing fields in the request message that are duplicated
by the blinded path - namely, cltv limit + delta and target node id
(implied by the introduction node). This change takes the approach
of disallowing setting of conflicting fields when a path is present,
rather than asking the user to map blinded path fields to the existing
api (eg, provide the blinded path's final delta in the existing field).

Note that we *do not* switch out the values from the blinded path
on the API level. This is intentional, because we don't want our API
to need to be aware of the logic involved in queries to blinded paths
- this is a layering violation, and puts us at risk of forgetting to
duplicate this logic across our multiple "path-related" APIs. Instead,
our rpc logic knows the bare minimum required for sensible validation,
specifically "when this field is set, these others should not be".
This allows us to isolate blinding-related logic in the router.