multi: add blinding point to payment descriptor / use in hop decryption
If there is a blinding point present in update_add_htlc's TLVs, add
it to our payment descriptor struct so that it can be passed through
to onion decryption. This code path will be used for nodes that are
relaying payments inside of the blinded route (ie, all nodes except
the introduction node).

There is one instance where we create an Add type payment descriptor
without a blinding point (even if one was originally present for the
htlc) in lnwallet/channel.go, because we do not persist the htlc's
tlvs. This is okay because htlcs persisted as a part of our commit
state have already used the blinding point, and no longer need it
for processing.

TODO(carla): If we restore from state, and don't have blinding point,
will we be able to tell whether htlcs that we failed back were part
of a blinded route?