htlcswitch: store blinding point for introduction and relay nodes
When we have hops that are within a blinded route, we need to also
store the blinding point along with the regular onion ephemeral key
to be able to properly encrypt errors for the source. This change
updates the error encrypter to store this blinding point.

Note: the encoding/decoding here requires that the caller knows
the type of the error encrypter provided (ie, sphinx, introduction
or relay) and initializes the SphinxErrorEncryptor struct accordingly
before calling Encode/Decode to handle this read properly. This is
_already_ a requirement so that we know the difference between a
sphinx and mock encrypter (we write our type in htlcswitch/circuit.go)
so we just re-use this setup for blinded routes.