[workaround]lnwallet: set remote update log blinding point from pending
This commit adds a workaround for the fact that we don't have our extra
htlc data (ie, that provided in tlvs) stored in channeldb.HTLC. It
ensures that we have the blinding point populated on restart for the
following set of circumstances:
* The incoming HTLC is irrevocably committed to in our local commitment,
  ie: we have received a CommitSig and sent a RevokeAndAck for a
  commitment that includes the incoming HTLC.
* The incoming HTLC is still pending on the sender, ie: we have not yet
  sent the remote party a CommitSig covering the incoming HTLC.

If we restart at this point, the htlc will be stored as an incoming htlc
on our local commitment (with no blinding point) and the full log update
will be saved as a LogUpdate in our remote pending updates (because we
have not yet provided the remote party with a signature).

We restore our remoteUpdateLog from the local commit's incoming htlcs,
so before this change these htlcs would be loaded into our in-memory log
without their blinding point set (and operation would incorrectly resume
as usual without it). This commit updates our logic to restore blinding
points (if set) to the htlcs in the remoteUpdateLog.