lnwallet: add custom records to payment descriptor
Where available, add custom records included in UpdateAddHTLC to
PaymentDescriptor. When restoring from disk, we do not have these
fields present as we don't want to store (possibly junk) data sent to
us by remote peers.

A possible extension to LND to allow storing this data is to allow
node operators to configure the TLV records that they would like to
store, though this is not supported in this change.