routing: add result interpretation for intermediate invalid blinding
This commit adds handling for route blinding errors that are reported
by the introduction node in a multi-hop blinded route. It penalizes
all nodes _after_ the reporting node, since we always expect the
introduction node to report errors for blinded routes. The introduction
node is not penalized to protect an individual node from being unfairly
penalized for reporting the errors experienced within a route.

If this error code is reported by a node that is not an introduction
node, we penalize the node because it is returning an error code that
it should not be using.