htlcswitch: switch out blinded failures for local errors
When we're the error origin of a HTLC failure that is part of a blinded
route, switch out the error to have an invalid onion blinding error
(per the route blinding specification). It doesn't matter whether we're
the introduction node or a relaying blinded hop here, because either
way the htlc that we were going to forward will have a non-nil blinding
point (set on the to-be-failed UpdateAddHTLC when it was created for
forwarding).

The specification does have special instructions for the receiving node,
but we don't need to worry about these because:
1. LND does not yet support blinded receives.
2. This code is only ever called for HTLC forwards, not receives.