htlcswitch: switch out remote failure for blinding error
This commit updates remote failures from our peers to use blinded
error failure codes is a blinding point is present on the outgoing
HTLC.

Note that in the case of receiving a malformed HTLC error, we may very
well convert our error to a UpdateFailHTLC just to convert it back to
a malformed error on the incoming link (because we are not the
introduction node). We choose to do this over upgrading the switch to
propagate malformed errors, as the functionality we require is still
that of a failure, we're just send our peer less information than usual
to protect the blinded route against probing.