htlcswitch: add hop type to invalid payload errors
In some cases where we create an invalid payload error, we've managed
to successfully parse the payload but it is logically invalid. If the
invalid payload is part of a blinded route, we still want to be able
to handle errors correctly, so we add a hop type to our errors so that
we can convert them accordingly. We don't _always_ parse the payload
when we return these errors, so we can't just check for a blinding
point at the error site. Adding a hop type helps to bubble this
information in in the error.