htlcswitch: switch out local link payload failures for blinding error
This commit updates the link to set the blinding code for payload
failures for relaying nodes. Note that this is the only case of an
incoming failure where we actually get to see whether we have a
blinding point in our payload (since other incoming failures are due
to a parsing issue with the payload).