htlcswitch: add ListOpenCircuits to circuit map
When we restart, we want to provide the ResourceManager with a view
of our currently in-flight HTLCs. We can do this by providing a list
operation on the circuit manager.

We choose to copy the PaymentCircuit in this API to prevent any
accidental mutation, as the structs themselves aren't big enough to
justify the risk.