congestion[checkme]: restore view of in flight HTLCs from circuitmap
This commit restores the current in flight view of HTLCs for the
resource manager on restart.

Things to check:
- Consistency between circuit map and channel state
- Are there any gaps where we'd miss a HTLC
- What about settles/fails, where do they pipe in
- Other stuff I can't think about now

Thoughts on timing: it doesn't seem terrible to _not_ have the added
time for our HTLCs. Maybe this will get added and that's nice, but if
we temporarily count in flight htlcs a little less on restart, it's
not dreadful dreadful (unless exploited along with a crashbug where an
adversary could time this exactly; to combat that, we could just assume
the absolute worst to over rather than under estimate).